fix: retry Agent suggestion send while another message is pending
Keep the composer prefill buffer when sendPending is true so install boot
kickoff cannot drop a selected improvement prompt after it is dismissed.
